# docs/openapi.yaml
components:
  schemas:
...
    CasesRequest:
      properties:
        fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          default:
          - case_id
          description: fields (optional)
          title: Fields
        filter:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: filter (optional)
          title: Filter
        offset:
          anyOf:
          - maximum: 10000.0
            minimum: 0.0
            type: integer
          - type: 'null'
          default: 0
          description: 'offset (optional) default: 0'
          title: Offset
        size:
          anyOf:
          - maximum: 10000.0
            minimum: 1.0
            type: integer
          - type: 'null'
          default: 10
          description: 'number of cases to return (optional) default: 10'
          title: Size
      title: CasesRequest
      type: object
    CohortQueryRequest:
      properties:
        case_ids_filter_path:
          description: path for the case ids in the query
          title: Case Ids Filter Path
          type: string
        case_index:
          anyOf:
          - type: string
          - type: 'null'
          default: CaseCentric_case_centric
          description: case index to query
          title: Case Index
        cohort_filter:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: case filter (optional)
          title: Cohort Filter
        cohort_item_field:
          anyOf:
          - type: string
          - type: 'null'
          default: case_id
          description: identity field for the case
          title: Cohort Item Field
        filter:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: query filter (optional)
          title: Filter
        limit:
          anyOf:
          - type: integer
          - type: 'null'
          default: 10000
          description: set the number of responses (optional)
          title: Limit
        query:
          anyOf:
          - type: string
          - type: 'null'
          default: ''
          description: the query to execute using the case ids found using the cohort_filter(optional)
          title: Query
      required:
      - case_ids_filter_path
      title: CohortQueryRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    TopGenesChartRequest:
      properties:
        cohort_filters:
          additionalProperties: true
          default:
            and: []
          title: Cohort Filters
          type: object
        size:
          default: 20
          title: Size
          type: integer
        ssm_filters:
          additionalProperties: true
          default:
            and:
            - nested:
                in:
                  project_id:
                  - MMRF-COMMPASS
                path: case.project
          title: Ssm Filters
          type: object
      title: TopGenesChartRequest
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  title: Gen3 Analysis Service
  version: 1.0.3
openapi: 3.1.0
paths:
...
  /cases/:
    post:
      description: Performs a cases query and returns case metadata for the matching
        cases.
      operationId: query_cases_cases__post
      parameters:
      - in: cookie
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CasesRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successfully processed the cohort query
        '400':
          description: The request body is missing required fields or has invalid
            values.
        '401':
          description: User unauthorized when accessing endpoint
        '403':
          description: User does not have access to requested data
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          description: Something went wrong internally when processing the request
      summary: Query case metadata
      tags:
      - Cases
  /cases/{case_id}:
    get:
      description: Performs a cases query and returns case metadata for case id
      operationId: get_case_by_id_cases__case_id__get
      parameters:
      - description: case id
        in: path
        name: case_id
        required: true
        schema:
          description: case id
          title: Case Id
          type: string
      - in: cookie
        name: access_token
        required: false
        schema:
          anyOf:
          - maxItems: 1
            minItems: 1
            prefixItems:
            - {}
            type: array
          - type: 'null'
          title: Access Token
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successfully processed the case query
        '400':
          description: The request body is missing required fields or has invalid
            values.
        '401':
          description: User unauthorized when accessing endpoint
        '403':
          description: User does not have access to requested data
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          description: Something went wrong internally when processing the request
      summary: Get case id metadata
      tags:
      - Cases
...
  /cohorts/query:
    post:
      description: Performs a cohort query and return the query for all items matching
        the ids.
      operationId: cohort_query_cohorts_query_post
      parameters:
      - in: cookie
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortQueryRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successfully processed the cohort query
        '400':
          description: The request body is missing required fields or has invalid
            values.
        '401':
          description: User unauthorized when accessing endpoint
        '403':
          description: User does not have access to requested data
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          description: Something went wrong internally when processing the request
      security:
      - HTTPBearer: []
      summary: Queries for cohort ids and uses those ids as the cohort in the second
        query
      tags:
      - Cohorts
  /cohorts/top_genes_in_cohort:
    post:
      operationId: top_genes_in_cohort_cohorts_top_genes_in_cohort_post
      parameters:
      - in: cookie
        name: access_token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopGenesChartRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Top Genes In Cohort
      tags:
      - Cohorts
...